Add files via upload
The code imports the remove function from the rembg module and the Image module from the Python Imaging Library (PIL). It opens an image file named "birds-pet-parrot.jpg" using the Image.open() method and stores it in the img variable. Then, the remove() function is called with img as the argument to remove the background of the image. The resulting image is stored in the R variable. Finally, the save() method is called on R to save the image as "img1.png".